refactor(web-terminal): split the popover's derived facts out
The redesigned popover crossed the interface-JS max-lines cap (470 of
450 counted lines). The pure derivations — refusal phrases, lock
reasons, the head note, the reachability wording — never touched the
DOM or the module state, so they move to control-target-facts.js
whole; the popover keeps the rows, gestures and confirms, and
re-exports lockReason so its surface is unchanged.
